@@ -14,10 +14,10 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^14.3.0",
-    "@angular/cdk": "^13.3.9",
+    "@angular/cdk": "^14.2.7",
     "@angular/common": "^14.3.0",
     "@angular/compiler": "^14.3.0",
-    "@angular/core": "^13.4.0",
+    "@angular/core": "^14.3.0",
     "@angular/forms": "^14.3.0",
     "@angular/localize": "^14.3.0",
     "@angular/platform-browser": "^14.3.0",
@@ -55,8 +55,8 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular/cli": "^13.3.11",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
     "@angular/compiler-cli": "^14.3.0",
     "@angular/language-service": "^14.3.0",
     "@types/jasmine": "~3.6.0",
