@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup, Validators} from '@angular/forms';
 import {PersonService} from '../../shared/services/person.service';
 import {ConferenceService} from '../../shared/services/conference.service';
 import {Router} from '@angular/router';
@@ -13,10 +13,10 @@ import * as _ from 'lodash';
 })
 export class ForgotPasswordComponent implements OnInit {
 
-  userForm: FormGroup;
+  userForm: UntypedFormGroup;
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private personSrv: PersonService,
     private conferenceSrv: ConferenceService,
     private router: Router,
